+++ src/use_event_listener.rs
@@ -1,5 +1,6 @@
 use crate::core::ElementMaybeSignal;
 use crate::{watch_with_options, WatchOptions};
+use default_struct_builder::DefaultBuilder;
 use leptos::ev::EventDescriptor;
 use leptos::*;
 use std::cell::RefCell;
@@ -17,11 +18,11 @@ use wasm_bindgen::JsCast;
 /// # use leptos::*;
 /// # use leptos::ev::visibilitychange;
 /// # use leptos::logging::log;
-/// # use leptos_use::use_event_listener;
+/// # use leptos_use::{use_document, use_event_listener};
 /// #
 /// # #[component]
 /// # fn Demo() -> impl IntoView {
-/// use_event_listener(document(), visibilitychange, |evt| {
+/// use_event_listener(use_document(), visibilitychange, |evt| {
 ///     log!("{:?}", evt);
 /// });
 /// #    view! { }
@@ -81,7 +82,7 @@ use wasm_bindgen::JsCast;
 ///
 /// ## Server-Side Rendering
 ///
-/// Please refer to ["Functions with Target Elements"](https://leptos-use.rs/server_side_rendering.html#functions-with-target-elements)
+/// On the server this amounts to a noop.
 pub fn use_event_listener<Ev, El, T, F>(target: El, event: Ev, handler: F) -> impl Fn() + Clone
 where
     Ev: EventDescriptor + 'static,
@@ -89,12 +90,7 @@ where
     T: Into<web_sys::EventTarget> + Clone + 'static,
     F: FnMut(<Ev as EventDescriptor>::EventType) + 'static,
 {
-    use_event_listener_with_options(
-        target,
-        event,
-        handler,
-        web_sys::AddEventListenerOptions::new(),
-    )
+    use_event_listener_with_options(target, event, handler, UseEventListenerOptions::default())
 }
 
 /// Version of [`use_event_listener`] that takes `web_sys::AddEventListenerOptions`. See the docs for [`use_event_listener`] for how to use.
@@ -102,7 +98,7 @@ pub fn use_event_listener_with_options<Ev, El, T, F>(
     target: El,
     event: Ev,
     handler: F,
-    options: web_sys::AddEventListenerOptions,
+    options: UseEventListenerOptions,
 ) -> impl Fn() + Clone
 where
     Ev: EventDescriptor + 'static,
@@ -115,7 +111,7 @@ where
 
     let cleanup_fn = {
         let closure_js = closure_js.clone();
-        let options = options.clone();
+        let options = options.as_add_event_listener_options();
 
         move |element: &web_sys::EventTarget| {
             let _ = element.remove_event_listener_with_callback_and_event_listener_options(
@@ -152,6 +148,8 @@ where
                 prev_element.replace(element.clone());
 
                 if let Some(element) = element {
+                    let options = options.as_add_event_listener_options();
+
                     _ = element.add_event_listener_with_callback_and_add_event_listener_options(
                         &event_name,
                         closure_js.as_ref().unchecked_ref(),
@@ -171,4 +169,53 @@ where
     on_cleanup(stop.clone());
 
     stop
+}
+
+/// Options for [`use_event_listener_with_options`].
+#[derive(DefaultBuilder, Default, Copy, Clone)]
+pub struct UseEventListenerOptions {
+    /// A boolean value indicating that events of this type will be dispatched to
+    /// the registered `listener` before being dispatched to any `EventTarget`
+    /// beneath it in the DOM tree. If not specified, defaults to `false`.
+    capture: bool,
+
+    /// A boolean value indicating that the `listener` should be invoked at most
+    /// once after being added. If `true`, the `listener` would be automatically
+    /// removed when invoked. If not specified, defaults to `false`.
+    once: bool,
+
+    /// A boolean value that, if `true`, indicates that the function specified by
+    /// `listener` will never call
+    /// [`preventDefault()`](https://developer.mozilla.org/en-US/docs/Web/API/Event/preventDefault "preventDefault()").
+    /// If a passive listener does call `preventDefault()`, the user agent will do
+    /// nothing other than generate a console warning. If not specified,
+    /// defaults to `false` – except that in browsers other than Safari,
+    /// defaults to `true` for the
+    /// [`wheel`](https://developer.mozilla.org/en-US/docs/Web/API/Element/wheel_event "wheel"),
+    /// [`mousewheel`](https://developer.mozilla.org/en-US/docs/Web/API/Element/mousewheel_event "mousewheel"),
+    /// [`touchstart`](https://developer.mozilla.org/en-US/docs/Web/API/Element/touchstart_event "touchstart") and
+    /// [`touchmove`](https://developer.mozilla.org/en-US/docs/Web/API/Element/touchmove_event "touchmove")
+    /// events. See [Improving scrolling performance with passive listeners](https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener#improving_scrolling_performance_with_passive_listeners)
+    /// to learn more.
+    #[builder(into)]
+    passive: Option<bool>,
+}
+
+impl UseEventListenerOptions {
+    fn as_add_event_listener_options(&self) -> web_sys::AddEventListenerOptions {
+        let UseEventListenerOptions {
+            capture,
+            once,
+            passive,
+        } = self;
+
+        let mut options = web_sys::AddEventListenerOptions::new();
+        options.capture(*capture);
+        options.once(*once);
+        if let Some(passive) = passive {
+            options.passive(*passive);
+        }
+
+        options
+    }
 }